+++ controllers/index_controllers.js
@@ -1,0 +1,7 @@
+module.exports.index = (req, res) => {
+    res.json({ 
+        message: 'quality anime quotes coming soon 🚀', 
+        error: false, 
+        data: null 
+    });
+}
